add service

+++ main.py
@@ -7,6 +7,8 @@ import models
 from database import *
 from sqlalchemy.orm import Session
 
+import myservice
+
 app = FastAPI()
 
 Base.metadata.create_all(engine)
@@ -17,6 +19,12 @@ def get_session():
     finally:
         session.close()
 
+# def create_default_config(session = Depends(get_session)):;
+#     item = models.Item(service = item.service, data = item.data);
+#     session.add(item);
+#     session.commit();
+#     session.refresh(item);
+# create_default_config();
 
 @app.get("/help")
 def help():
@@ -44,8 +52,22 @@ def update_item(id:int, item:schemas.Item, session = Depends(get_session)):
 
 @app.delete("/{id}")
 def delete_item(id:int, session = Depends(get_session)):
+    if myservice.use_config() == id:
+        return "You can not delete the config used by the service."
     itemObject = session.query(models.Item).get(id)
     session.delete(itemObject)
     session.commit()
     session.close()
     return f"Config id={id} delete."
+
+@app.get("/service")
+def get_service_config(session: Session = Depends(get_session)):
+    item = None
+    if myservice.use_config():
+        item = session.query(models.Item).get(myservice.use_config())
+    return item
+
+@app.put("/service/{id}")
+def update_service_config(id:int, session: Session = Depends(get_session)):
+    item = session.query(models.Item).get(myservice.use_config(id))
+    return "Config update!", item
